+++ packages/components/pie-select/test/helpers/page-object/selectors.ts
@@ -1,0 +1,13 @@
+export const select = {
+    selectors: {
+        container: {
+            dataTestId: 'pie-select-shell',
+        },
+        select: {
+            dataTestId: 'pie-select-element',
+        },
+        assistiveText: {
+            dataTestId: 'pie-select-assistive-text',
+        },
+    },
+};